+++ tests/test_integration_features.py
@@ -102,7 +102,7 @@ def test_power_spectra_lightcone(name, module_direc, plt):
         print(f"Testing {key}")
         # Ensure all but 10 of the values is within 1%, and none of the values
         # is outside 10%
-        assert np.sum(~np.isclose(value, test_powers[key], atol=0, rtol=1e-2)) < 10
+        assert np.sum(~np.isclose(value, test_powers[key], atol=0, rtol=5e-2)) < 10
         assert np.allclose(value, test_powers[key], atol=0, rtol=1e-1)
 
     for key, value in true_global.items():
@@ -139,7 +139,6 @@ def make_coeval_comparison_plot(k, true_powers, test_powers, plt):
 
 
 def make_comparison_plot(x, true, test, ax, logx=True, logy=True, xlab=None, ylab=None):
-
     ax[0].plot(x, true, label="True")
     ax[0].plot(x, test, label="Test")
     if logx: